Part 3: Optional: Dockerize the API
Added a dockerfile, yaml, file and docker compose file that dockerize the API for easy deployment and portability.
Moreover, added images showing the docker build command output and updated the necessary libraries for the API into a requirements.txt file used by the container. And updated readme to include the detailed dockerization steps and explanations